+++ components/Container/__tests__/index.spec.tsx
@@ -125,18 +125,13 @@ describe('SidebarPanel', () => {
     expect(sidebar).toHaveAttribute('id', 'customID');
   });
 
-  it('can toggel open and closed state', () => {
+  it('should set state-open class on SidebarPanel', () => {
     const { rerender } = render(<SidebarPanel isOpen={true}><h1>test</h1></SidebarPanel>);
 
     const sidebar = screen.getByTestId('sidebar-panel');
 
     expect(sidebar).toHaveClass('state-open');
     expect(sidebar).not.toHaveClass('state-closed');
-
-    rerender(<SidebarPanel isOpen={false}><h1>test</h1></SidebarPanel>);
-
-    expect(sidebar).toHaveClass('state-closed');
-    expect(sidebar).not.toHaveClass('state-open');
   });
 });