+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Added
+- Added the `Question details` page that allows users to answer a question [#320]
+- Added, or moved, the `question details` page from `projects/[projectId]/dmp/[dmpid]/q/[qid]` to `projects/[projectId]/dmp/[dmpid]/s/[sid]/q/[qid]` [#320]
+- Added a new `useRenderQuestionField` hook to render different questions on the `Question details` page [#320]
 - Added delete section functionality to template builder with confirmation modal, including translations, and tests
 - Implemented a "Delete Question" feature on the question editing page with extra dialog and with tests